requirements.txt
# requirements.txt - 必要なPythonパッケージ一覧
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.1.0
webdriver-manager>=3.8.0
pandas>=1.4.0
python-dotenv>=0.20.0
google-api-python-client>=2.47.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.1
schedule>=1.1.0
psycopg2-binary>=2.9.3

# プラットフォーム関連
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.1.0
webdriver-manager>=3.8.0
pandas>=1.4.0
python-dotenv>=0.20.0
google-api-python-client>=2.47.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.1
schedule>=1.1.0
psycopg2-binary>=2.9.3

# 追加機能用
Flask>=2.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.22.0
Pillow>=9.0.0
openai>=0.27.0
gunicorn>=20.1.0